Prepared for evaluation

+++ ft_substr.c
@@ -6,7 +6,7 @@
 /*   By: mbonengl <mbonengl@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/07 21:55:30 by mbonengl          #+#    #+#             */
-/*   Updated: 2024/06/18 20:03:29 by mbonengl         ###   ########.fr       */
+/*   Updated: 2024/06/21 18:58:54 by mbonengl         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,15 +25,15 @@ char	*alloc_zero(void)
 char	*ft_substr(char const *s, unsigned int start, size_t len)
 {
 	char	*substring;
-	int		s_len;
+	size_t	s_len;
 
 	if (!s)
 		return (NULL);
 	s_len = ft_strlen((char *)s);
 	if (start > (unsigned int)s_len)
 		return (alloc_zero());
-	if (len > ft_strlen((char *)s + start))
-		len = ft_strlen((char *)s + start);
+	if (len > (size_t)ft_strlen((char *)s + start))
+		len = (size_t)ft_strlen((char *)s + start);
 	substring = (char *)ft_calloc((len + 1), sizeof(char));
 	if (!substring)
 		return (NULL);
